Refactor fs2-stream
