+++ akka-stream/src/main/scala/neotypes/akkastreams/package.scala
@@ -1,0 +1,9 @@
+package neotypes
+
+import akka.stream.scaladsl.Source
+
+import scala.concurrent.Future
+
+package object akkastreams {
+  type AkkaStream[T] = Source[T, Future[Unit]]
+}
