+++ cats-effect/src/main/scala/neotypes/cats/implicits.scala
@@ -1,9 +1,7 @@
 package neotypes.cats
 
-import cats.effect
-
 object implicits {
-  implicit def catsAsync[F[_]](implicit F: effect.Async[F]): neotypes.Async[F] =
+  implicit def catsAsync[F[_]](implicit F: cats.effect.Async[F]): neotypes.Async[F] =
     new neotypes.Async[F] {
       override def async[T](cb: (Either[Throwable, T] => Unit) => Unit): F[T] =
         F.async(cb)
