+++ fs2-stream/src/test/scala/neotypes/fs2/Fs2StreamSpec.scala
@@ -8,17 +8,15 @@ import neotypes.implicits._
 import org.scalatest.AsyncFlatSpec
 
 class Fs2StreamSpec extends AsyncFlatSpec with BaseIntegrationSpec {
-  type Fs2Stream[T] = fs2.Stream[IO, T]
-
   it should "work with fs2 using cats.effect.IO" in {
     val s = driver.session().asScala[IO]
 
     "match (p:Person) return p.name"
       .query[Int]
-      .stream[Fs2Stream, IO](s)
+      .stream[Fs2IoStream, IO](s)
       .compile
       .toList
-      .unsafeToFuture
+      .unsafeToFuture()
       .map {
         names => assert(names == (0 to 10).toList)
       }
